Allow user input to change number of cells in grid
Add a button on the top of the screen that will
send the user a popup asking for the number of
squares per side for the new grid.

Once entered, the existing grid will be removed,
and a new grid will be generated in the same
total space as before so that you've got a
new sketch pad.